requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pymongo==4.6.1

